# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Reservation:
    properties:
      endTime:
        type: string
      lotID:
        type: integer
      reservationID:
        type: integer
      spaceID:
        type: integer
      startTime:
        type: string
      status:
        type: string
      vehicleID:
        type: integer
    type: object
  models.User:
    properties:
      password:
        type: string
      tel:
        type: string
      userID:
        type: integer
      userName:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: 1351688749@qq.com
    name: API Support
    url: http://example.com/support
  description: API documentation for Parking Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Parking Management API
  version: "1.0"
paths:
...
  /api/reservation:
    post:
      consumes:
      - application/json
      description: Create a new reservation
      parameters:
      - description: Reservation Details
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/models.Reservation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a reservation
      tags:
      - Reservation
  /api/reservation/{id}:
    delete:
      consumes:
      - application/json
      description: Cancel an existing reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancel a reservation
      tags:
      - Reservation
    put:
      consumes:
      - application/json
      description: Update the details of an existing reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Reservation Details
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/models.Reservation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update reservation details
      tags:
      - Reservation
  /api/reservation/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an existing reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Reservation Details
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/models.Reservation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update reservation status
      tags:
      - Reservation
  /api/reservation/fee/{id}:
    get:
      consumes:
      - application/json
      description: Get the fee for a reservation based on lot ID and time range
      parameters:
      - description: Lot ID
        in: path
        name: id
        required: true
        type: integer
      - description: Start Time
        in: query
        name: start
        required: true
        type: string
      - description: End Time
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get fee for a reservation
      tags:
      - Reservation
  /api/user/authenticated:
    get:
      consumes:
      - application/json
      description: Get authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get authenticated user information
      tags:
      - User
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Login a user with given credentials
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login a user
      tags:
      - User
  /api/user/logout:
    post:
      consumes:
      - application/json
      description: Logout a user by clearing the JWT cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Logout a user
      tags:
      - User
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Register a new user with given parameters
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register a new user
      tags:
      - User
...
swagger: "2.0"